kvx: eth: sanity check for ring mgmt
Summary:
Add sanity checks for ring management

Ref T19567

Test Plan: eth*

Reviewers: O39 ethernet!

Maniphest Tasks: T19567

Differential Revision: https://phab.kalray.eu/D25069

+++ drivers/net/ethernet/kalray/kvx-net.h
@@ -100,9 +100,9 @@ struct kvx_eth_ring {
 	};
 	struct napi_struct napi;
 	struct sk_buff *skb;
-	u16 count;
-	u16 next_to_use;
-	u16 next_to_clean;
+	unsigned int count;
+	unsigned int next_to_use;
+	unsigned int next_to_clean;
 	int qidx;
 	bool init_done;
 	enum rx_ring_type type;
